requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
filelock==3.15.1
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2024.2
uvicorn==0.30.6
virtualenv==20.26.2
whitenoise==6.8.2
